+++ src/main/kotlin/study/racinggame/service/reader/ActualInputReader.kt
@@ -1,4 +1,4 @@
-package study.racinggame.race.reader
+package study.racinggame.service.reader
 
 class ActualInputReader : InputReader {
     override fun raedLine() = readln()
